requirements.txt
advent-of-code-data==2.0.4
aocd-example-parser==2023.12.20
beautifulsoup4==4.12.3
charset-normalizer==3.4.0
idna==3.10
numpy==2.1.3
pandas==2.2.3
Pebble==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.2.3
wheel==0.44.0
